+++ core-ng/src/test/java/core/framework/impl/web/route/PathPatternValidatorTest.java
@@ -10,27 +10,34 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 class PathPatternValidatorTest {
     @Test
     void validate() {
-        new PathPatternValidator("/robot.txt").validate();
-        new PathPatternValidator("/images").validate();
+        new PathPatternValidator("/robot.txt", true).validate();
+        new PathPatternValidator("/images", true).validate();
 
-        new PathPatternValidator("/path-with-trailing-slash/").validate();
+        new PathPatternValidator("/path-with-trailing-slash/", true).validate();
 
-        new PathPatternValidator("/user/:id/name").validate();
-        new PathPatternValidator("/v2/user/:id").validate();
-        new PathPatternValidator("/ajax/:path(*)").validate();
+        new PathPatternValidator("/user/:id/name", true).validate();
+        new PathPatternValidator("/v2/user/:id", true).validate();
+        new PathPatternValidator("/ajax/:path(*)", true).validate();
     }
 
     @Test
     void duplicateVariable() {
-        assertThatThrownBy(() -> new PathPatternValidator("/:name/path/:name").validate())
+        assertThatThrownBy(() -> new PathPatternValidator("/:name/path/:name", true).validate())
                 .isInstanceOf(Error.class)
                 .hasMessageContaining("duplicate");
     }
 
     @Test
     void invalidVariable() {
-        assertThatThrownBy(() -> new PathPatternValidator("/path/:name(").validate())
+        assertThatThrownBy(() -> new PathPatternValidator("/path/:name(", true).validate())
                 .isInstanceOf(Error.class)
                 .hasMessageContaining(":name(");
+    }
+
+    @Test
+    void notAllowWildcardVariable() {
+        assertThatThrownBy(() -> new PathPatternValidator("/path/:name(*)", false).validate())
+                .isInstanceOf(Error.class)
+                .hasMessageContaining("wildcard path variable is not allowed");
     }
 }